#4885: Move program device map to program, no more need to explicitly clear program buffer cache, since gets cleared on program destructor
